Move search outside of src dir
