+++ search/bin/query.ts
