+++ search/bin/upsert-vectors.ts
