+++ search/mdxToPlainText.ts
