+++ search/package.json
@@ -1,0 +1,83 @@
+{
+  "name": "@zaduma/semantic-search",
+  "version": "0.0.1",
+  "private": true,
+  "packageManager": "pnpm@7.25.0",
+  "scripts": {},
+  "dependencies": {
+    "@astrojs/image": "^0.12.1",
+    "@astrojs/markdown-remark": "^1.1.3",
+    "@astrojs/mdx": "^0.14.0",
+    "@astrojs/rss": "^1.1.0",
+    "@astrojs/solid-js": "^1.2.3",
+    "@astrojs/tailwind": "^2.1.3",
+    "@fontsource/brygada-1918": "^4.5.11",
+    "@fontsource/inter": "^4.5.14",
+    "@mdx-js/mdx": "2.2.1",
+    "@types/lz-string": "^1.3.34",
+    "@types/react": "^18.0.26",
+    "@vercel/og": "^0.0.21",
+    "astro": "^1.9.2",
+    "astro-eslint-parser": "^0.9.5",
+    "eslint-plugin-markdown": "^3.0.0",
+    "estree-util-value-to-estree": "^2.0.0",
+    "gpt3-tokenizer": "^1.1.4",
+    "lite-youtube-embed": "^0.2.0",
+    "plaiceholder": "^2.5.0",
+    "postcss": "8.4.20",
+    "reading-time": "^1.5.0",
+    "remark-shiki-twoslash": "^3.1.0",
+    "remark-supersub": "^1.0.0",
+    "sharp": "^0.31.3",
+    "solid-js": "^1.4.3",
+    "tailwindcss": "^3.2.4",
+    "unified": "^10.1.2",
+    "unist-util-visit": "^4.1.1",
+    "vercel": "^28.10.1",
+    "vitest": "^0.26.3",
+    "yargs-parser": "^21.1.1"
+  },
+  "devDependencies": {
+    "@astrojs/rss": "^2.0.0",
+    "@edgeandnode/eslint-config": "^1.0.3",
+    "@types/hast": "2.3.4",
+    "@types/node": "^18.11.18",
+    "autoprefixer": "^10.4.13",
+    "cssnano": "^5.1.14",
+    "dotenv": "^16.0.3",
+    "esbuild-runner": "^2.2.2",
+    "eslint": "^8.30.0",
+    "node-fetch": "^3.3.0",
+    "openai": "^3.1.0",
+    "pinecone-client": "^1.0.1",
+    "prettier": "^2.8.1",
+    "prettier-plugin-astro": "^0.7.0",
+    "prettier-plugin-tailwindcss": "^0.2.1",
+    "remark": "^14.0.2",
+    "remark-mdx": "^2.2.1",
+    "remark-mdx-to-plain-text": "^3.0.0",
+    "typescript": "^4.9.4",
+    "unist-util-remove": "^3.1.0"
+  },
+  "pnpm": {
+    "peerDependencyRules": {
+      "ignoreMissing": [
+        "@babel/core",
+        "rollup",
+        "esbuild"
+      ]
+    },
+    "overrides": {
+      "@mdx-js/mdx": "2.2.1",
+      "typescript": "^4.9.4",
+      "@types/node": "18.11.18",
+      "esbuild": "0.15.18"
+    },
+    "patchedDependencies": {
+      "@mdx-js/mdx@2.2.1": "patches/@mdx-js__mdx@2.1.3.patch"
+    }
+  },
+  "volta": {
+    "node": "18.13.0"
+  }
+}
