--- src/pages/search.astro
@@ -1,23 +1,0 @@
-<html>
-  <body>
-    <input id="input-id" />
-    <button data-callback="onSubmit" onclick="onSubmit" data-action="submit">
-      Click me to verify the captcha challenge!
-    </button>
-
-    <script is:inline>
-      function onSubmit(token) {
-        console.log("onSubmit");
-        const query = document.getElementById("input-id").value;
-        fetch("/api/search", {
-          method: "POST",
-          body: JSON.stringify({ query }),
-        })
-          .then((response) => response.json())
-          .then((gResponse) => {
-            console.log(gResponse);
-          });
-      }
-    </script>
-  </body>
-</html>
